requirement.txt
transformers
torch
gradio
newspaper3k